// src/proto/jobconfigpb/resolvable_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file jobconfigpb/resolvable.proto (package dev.reduction.jobconfig, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file jobconfigpb/resolvable.proto.
 */
export const file_jobconfigpb_resolvable: GenFile = /*@__PURE__*/
  fileDesc("Chxqb2Jjb25maWdwYi9yZXNvbHZhYmxlLnByb3RvEhdkZXYucmVkdWN0aW9uLmpvYmNvbmZpZyI1CglTdHJpbmdWYXISDwoFdmFsdWUYASABKAlIABIPCgVwYXJhbRgCIAEoCUgAQgYKBGtpbmQiNAoISW50MzJWYXISDwoFdmFsdWUYASABKAVIABIPCgVwYXJhbRgCIAEoCUgAQgYKBGtpbmRCLloscmVkdWN0aW9uLmRldi9yZWR1Y3Rpb24tcHJvdG9jb2wvam9iY29uZmlncGJiBnByb3RvMw");

/**
 * @generated from message dev.reduction.jobconfig.StringVar
 */
export type StringVar = Message<"dev.reduction.jobconfig.StringVar"> & {
  /**
   * @generated from oneof dev.reduction.jobconfig.StringVar.kind
   */
  kind: {
    /**
     * @generated from field: string value = 1;
     */
    value: string;
    case: "value";
  } | {
    /**
     * @generated from field: string param = 2;
     */
    value: string;
    case: "param";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message dev.reduction.jobconfig.StringVar.
 * Use `create(StringVarSchema)` to create a new message.
 */
export const StringVarSchema: GenMessage<StringVar> = /*@__PURE__*/
  messageDesc(file_jobconfigpb_resolvable, 0);

/**
 * @generated from message dev.reduction.jobconfig.Int32Var
 */
export type Int32Var = Message<"dev.reduction.jobconfig.Int32Var"> & {
  /**
   * @generated from oneof dev.reduction.jobconfig.Int32Var.kind
   */
  kind: {
    /**
     * @generated from field: int32 value = 1;
     */
    value: number;
    case: "value";
  } | {
    /**
     * @generated from field: string param = 2;
     */
    value: string;
    case: "param";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message dev.reduction.jobconfig.Int32Var.
 * Use `create(Int32VarSchema)` to create a new message.
 */
export const Int32VarSchema: GenMessage<Int32Var> = /*@__PURE__*/
  messageDesc(file_jobconfigpb_resolvable, 1);